Ignore some play tests because they are using javascript repository
