+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/RepoScriptBlockUtil.groovy
@@ -76,7 +76,6 @@ class RepoScriptBlockUtil {
             repositories.maven { MavenArtifactRepository repo ->
                 repo.name = name
                 repo.url = mirrorUrl
-                System.out.println("name: ${repo.name}, url: ${repo.url}")
             }
         }
     }
