+++ subprojects/jacoco/jacoco.gradle
@@ -22,6 +22,7 @@ dependencies {
     compile project(':plugins')
     compile project(':reporting')
     testCompile testLibraries.jsoup
+    testCompile project(':internalIntegTesting')
 }
 
 gradlebuildJava {
