+++ subprojects/jacoco/src/test/groovy/org/gradle/testing/jacoco/plugins/JacocoPluginSpec.groovy
@@ -17,8 +17,13 @@ package org.gradle.testing.jacoco.plugins
 
 import org.gradle.api.tasks.JavaExec
 import org.gradle.api.tasks.testing.Test
+import org.gradle.integtests.fixtures.RepoScriptBlockUtil
 import org.gradle.language.base.plugins.LifecycleBasePlugin
 import org.gradle.test.fixtures.AbstractProjectBuilderSpec
+import org.gradle.util.Requires
+import org.gradle.util.TestPrecondition
+import spock.lang.Issue
+import spock.lang.Unroll
 
 class JacocoPluginSpec extends AbstractProjectBuilderSpec {
     def setup() {
@@ -39,6 +44,58 @@ class JacocoPluginSpec extends AbstractProjectBuilderSpec {
         Test task = project.tasks.create('customTest', Test)
         expect:
         task.extensions.getByType(JacocoTaskExtension) != null
+    }
+
+    @Requires(TestPrecondition.ONLINE)
+    @Unroll
+    @Issue("GRADLE-3498")
+    def 'jacoco task extension can be configured. includeNoLocationClasses: #includeNoLocationClassesValue'() {
+        given:
+        project.apply plugin: 'java'
+        RepoScriptBlockUtil.configureJcenter(project.repositories)
+        def testTask = project.tasks.getByName('test')
+        JacocoTaskExtension extension = testTask.extensions.getByType(JacocoTaskExtension)
+
+        when:
+        extension.with {
+            destinationFile = project.file('build/jacoco/fake.exec')
+            append = false
+            includes = ['org.*', '*.?acoco*']
+            excludes = ['org.?joberstar']
+            excludeClassLoaders = ['com.sun.*', 'org.fak?.*']
+            includeNoLocationClasses = includeNoLocationClassesValue
+            sessionId = 'testSession'
+            dumpOnExit = false
+            output = JacocoTaskExtension.Output.TCP_SERVER
+            address = '1.1.1.1'
+            port = 100
+            classDumpDir = project.file('build/jacoco-dump')
+            jmx = true
+        }
+
+        def expected = new StringBuilder().with { builder ->
+            builder << "destfile=build/jacoco/fake.exec,"
+            builder << "append=false,"
+            builder << "includes=org.*:*.?acoco*,"
+            builder << "excludes=org.?joberstar,"
+            builder << "exclclassloader=com.sun.*:org.fak?.*,"
+            builder << "inclnolocationclasses=$includeNoLocationClassesValue,"
+            builder << "sessionid=testSession,"
+            builder << "dumponexit=false,"
+            builder << "output=tcpserver,"
+            builder << "address=1.1.1.1,"
+            builder << "port=100,"
+            builder << "classdumpdir=build/jacoco-dump,"
+            builder << "jmx=true"
+            builder.toString()
+        }
+
+        then:
+        def jvmArg = extension.asJvmArg
+        jvmArg.replaceFirst(/-javaagent:[^=]*\.jar=/, '') == expected
+
+        where:
+        includeNoLocationClassesValue << [true, false]
     }
 
     def "declares task property values for group and description"() {
