+++ subprojects/platform-play/src/integTest/groovy/org/gradle/play/integtest/continuous/PlayContinuousBuildReloadWaitingIntegrationTest.groovy
@@ -16,10 +16,12 @@
 
 package org.gradle.play.integtest.continuous
 
+import spock.lang.Ignore
 import spock.lang.Unroll
 /**
  * Test that app requests block while a build is in progress when using `--continuous`.
  */
+@Ignore('contains javascript repository')
 class PlayContinuousBuildReloadWaitingIntegrationTest extends AbstractPlayReloadIntegrationTest {
 
     def setup() {
