+++ subprojects/platform-play/src/integTest/groovy/org/gradle/play/integtest/continuous/PlayReloadIntegrationTest.groovy
@@ -16,9 +16,12 @@
 
 package org.gradle.play.integtest.continuous
 
+import spock.lang.Ignore
+
 /**
  * Test Play reload without `--continuous`
  */
+@Ignore('contains javascript repository')
 class PlayReloadIntegrationTest extends PlayContinuousBuildReloadIntegrationTest {
 
     def setup() {
